Unify send event into single command pd-send.py
